Add test utility classes for database integration tests
