+++ ynab-importer/infrastructure-it/src/test/scala/works/iterative/incubator/transactions/infrastructure/PostgreSQLSourceAccountRepositorySpec.scala
@@ -3,84 +3,21 @@ package infrastructure
 
 import zio.*
 import zio.test.*
-import zio.test.Assertion.*
 import com.augustnagro.magnum.magzio.*
-import com.dimafeng.testcontainers.PostgreSQLContainer
-import org.testcontainers.utility.DockerImageName
-import javax.sql.DataSource
-import com.zaxxer.hikari.{HikariConfig, HikariDataSource}
 import service.SourceAccountRepository
 import zio.test.TestAspect.sequential
 
 object PostgreSQLSourceAccountRepositorySpec extends ZIOSpecDefault:
-    private val postgresImage = DockerImageName.parse("postgres:17-alpine")
-
-    // Define a layer for test container
-    val postgresContainer =
-        ZLayer {
-            ZIO.acquireRelease {
-                ZIO.attempt {
-                    val container = PostgreSQLContainer(postgresImage)
-                    container.start()
-                    container
-                }
-            }(container =>
-                ZIO.attempt {
-                    container.stop()
-                }.orDie
-            )
-        }
-
-    // Create a DataSource from the container
-    val dataSourceLayer =
-        postgresContainer >>> ZLayer.fromZIO {
-            for
-                container <- ZIO.service[PostgreSQLContainer]
-                _ <- ZIO.attempt(Class.forName("org.postgresql.Driver"))
-                dataSource <- ZIO.acquireRelease(ZIO.attempt {
-                    val config = new HikariConfig()
-                    config.setJdbcUrl(container.jdbcUrl)
-                    config.setUsername(container.username)
-                    config.setPassword(container.password)
-                    config.setMaximumPoolSize(5)
-                    new HikariDataSource(config)
-                })(dataSource => ZIO.attempt(dataSource.close()).orDie)
-            yield dataSource
-        }
-
-    // Create a Transactor from the DataSource
-    val transactorLayer =
-        dataSourceLayer.flatMap(env => Transactor.layer(env.get[DataSource]))
+    import PostgreSQLLayers.*
+    import MigrateAspects.*
 
     // Setup the repository layer for testing
     val repositoryLayer =
-        transactorLayer >+> ZLayer {
+        transactorLayer >>> ZLayer {
             for
                 transactor <- ZIO.service[Transactor]
             yield PostgreSQLSourceAccountRepository(transactor)
         }
-
-    // Create PostgreSQLDataSource from DataSource for Flyway
-    val postgreSQLDataSourceLayer = dataSourceLayer >>> ZLayer {
-        ZIO.service[DataSource].map(ds => PostgreSQLDataSource(ds))
-    }
-
-    // Create FlywayMigrationService layer
-    val flywayMigrationServiceLayer = postgreSQLDataSourceLayer >>> FlywayMigrationService.layer
-
-    // Use Flyway to manage schema setup and teardown for tests
-    val setupDbSchema = ZIO.scoped {
-        for
-            // Get the migration service
-            migrationService <-
-                ZIO.service[FlywayMigrationService].provideSome[Scope](flywayMigrationServiceLayer)
-            // First clean the database to ensure a fresh state
-            _ <- migrationService.clean()
-            // Then run migrations to set up the schema
-            result <- migrationService.migrate()
-            _ <- ZIO.log(s"Migration complete: ${result.migrationsExecuted} migrations executed")
-        yield ()
-    }
 
     // Create sample source account for testing
     def createSampleSourceAccount(id: Long = 1): SourceAccount =
@@ -95,90 +32,85 @@ object PostgreSQLSourceAccountRepositorySpec extends ZIOSpecDefault:
             lastSyncTime = None
         )
 
-    def spec = suite("PostgreSQLSourceAccountRepository")(
-        test("should save and retrieve a source account") {
-            for
-                // Setup
-                _ <- setupDbSchema
-                repository <- ZIO.service[SourceAccountRepository]
-                sourceAccount = createSampleSourceAccount()
+    def spec = {
+        suite("PostgreSQLSourceAccountRepository")(
+            test("should save and retrieve a source account") {
+                for
+                    repository <- ZIO.service[SourceAccountRepository]
+                    sourceAccount = createSampleSourceAccount()
 
-                // Execute
-                _ <- repository.save(sourceAccount.id, sourceAccount)
-                retrieved <- repository.load(sourceAccount.id)
-            yield
-            // Assert
-            assertTrue(
-                retrieved.isDefined,
-                retrieved.get.id == sourceAccount.id,
-                retrieved.get.accountId == sourceAccount.accountId,
-                retrieved.get.bankId == sourceAccount.bankId
-            )
-        },
-        test("should find source accounts by query") {
-            for
-                // Setup
-                _ <- setupDbSchema
-                repository <- ZIO.service[SourceAccountRepository]
-                sourceAccount1 = createSampleSourceAccount(1)
-                sourceAccount2 = createSampleSourceAccount(2).copy(accountId = "987654321")
+                    // Execute
+                    _ <- repository.save(sourceAccount.id, sourceAccount)
+                    retrieved <- repository.load(sourceAccount.id)
+                yield
+                // Assert
+                assertTrue(
+                    retrieved.isDefined,
+                    retrieved.get.id == sourceAccount.id,
+                    retrieved.get.accountId == sourceAccount.accountId,
+                    retrieved.get.bankId == sourceAccount.bankId
+                )
+            },
+            test("should find source accounts by query") {
+                for
+                    repository <- ZIO.service[SourceAccountRepository]
+                    sourceAccount1 = createSampleSourceAccount(1)
+                    sourceAccount2 = createSampleSourceAccount(2).copy(accountId = "987654321")
 
-                // Execute
-                _ <- repository.save(sourceAccount1.id, sourceAccount1)
-                _ <- repository.save(sourceAccount2.id, sourceAccount2)
-                results <- repository.find(SourceAccountQuery())
-            yield
-            // Assert
-            assertTrue(
-                results.size == 2,
-                results.exists(a => a.id == sourceAccount1.id),
-                results.exists(a => a.id == sourceAccount2.id)
-            )
-        },
-        test("should find source accounts by filter") {
-            for
-                // Setup
-                _ <- setupDbSchema
-                repository <- ZIO.service[SourceAccountRepository]
-                sourceAccount1 = createSampleSourceAccount(1)
-                sourceAccount2 = createSampleSourceAccount(2).copy(accountId = "987654321")
+                    // Execute
+                    _ <- repository.save(sourceAccount1.id, sourceAccount1)
+                    _ <- repository.save(sourceAccount2.id, sourceAccount2)
+                    results <- repository.find(SourceAccountQuery())
+                yield
+                // Assert
+                assertTrue(
+                    results.size == 2,
+                    results.exists(a => a.id == sourceAccount1.id),
+                    results.exists(a => a.id == sourceAccount2.id)
+                )
+            },
+            test("should find source accounts by filter") {
+                for
+                    repository <- ZIO.service[SourceAccountRepository]
+                    sourceAccount1 = createSampleSourceAccount(1)
+                    sourceAccount2 = createSampleSourceAccount(2).copy(accountId = "987654321")
 
-                // Execute
-                _ <- repository.save(sourceAccount1.id, sourceAccount1)
-                _ <- repository.save(sourceAccount2.id, sourceAccount2)
-                results <- repository.find(SourceAccountQuery(accountId = Some("987654321")))
-            yield
-            // Assert
-            assertTrue(
-                results.size == 1,
-                results.head.id == sourceAccount2.id,
-                results.head.accountId == "987654321"
-            )
-        },
-        test("should update an existing source account") {
-            for
-                // Setup
-                _ <- setupDbSchema
-                repository <- ZIO.service[SourceAccountRepository]
-                sourceAccount = createSampleSourceAccount()
+                    // Execute
+                    _ <- repository.save(sourceAccount1.id, sourceAccount1)
+                    _ <- repository.save(sourceAccount2.id, sourceAccount2)
+                    results <- repository.find(SourceAccountQuery(accountId = Some("987654321")))
+                yield
+                // Assert
+                assertTrue(
+                    results.size == 1,
+                    results.head.id == sourceAccount2.id,
+                    results.head.accountId == "987654321"
+                )
+            },
+            test("should update an existing source account") {
+                for
+                    repository <- ZIO.service[SourceAccountRepository]
+                    sourceAccount = createSampleSourceAccount()
 
-                // Execute - first save
-                _ <- repository.save(sourceAccount.id, sourceAccount)
+                    // Execute - first save
+                    _ <- repository.save(sourceAccount.id, sourceAccount)
 
-                // Update the source account
-                updatedSourceAccount = sourceAccount.copy(bankId = "0100")
+                    // Update the source account
+                    updatedSourceAccount = sourceAccount.copy(bankId = "0100")
 
-                // Save the updated source account and retrieve
-                _ <- repository.save(sourceAccount.id, updatedSourceAccount)
-                retrieved <- repository.load(sourceAccount.id)
-            yield
-            // Assert
-            assertTrue(
-                retrieved.isDefined,
-                retrieved.get.bankId == "0100"
-            )
-        }
-    ).provideSomeShared[Scope](
+                    // Save the updated source account and retrieve
+                    _ <- repository.save(sourceAccount.id, updatedSourceAccount)
+                    retrieved <- repository.load(sourceAccount.id)
+                yield
+                // Assert
+                assertTrue(
+                    retrieved.isDefined,
+                    retrieved.get.bankId == "0100"
+                )
+            }
+        ) @@ sequential @@ migrate
+    }.provideSomeShared[Scope](
+        flywayMigrationServiceLayer,
         repositoryLayer
-    ) @@ sequential
+    )
 end PostgreSQLSourceAccountRepositorySpec
