+++ ynab-importer/infrastructure/src/main/scala/works/iterative/incubator/transactions/infrastructure/FlywayMigrationService.scala
@@ -17,9 +17,12 @@ end FlywayMigrationService
   *
   * @param locations
   *   Classpath locations to scan for migrations
+  * @param cleanDisabled
+  *   Whether database cleaning is disabled (true by default for safety)
   */
 case class FlywayConfig(
-    locations: List[String] = List(FlywayConfig.DefaultLocation)
+    locations: List[String] = List(FlywayConfig.DefaultLocation),
+    cleanDisabled: Boolean = true
 )
 
 object FlywayConfig:
@@ -40,6 +43,7 @@ class PostgreSQLFlywayMigrationService(
         val flywayConfig = Flyway.configure()
             .dataSource(dataSource.dataSource)
             .locations(config.locations*)
+            .cleanDisabled(config.cleanDisabled)
 
         flywayConfig.load()
     end flyway
