+++ ynab-importer/web/src/main/scala/works/iterative/incubator/transactions/infrastructure/TransactionImportModule.scala
@@ -32,8 +32,8 @@ import org.http4s.Uri
   */
 class TransactionImportModule(appShell: ScalatagsAppShell)
     extends ZIOWebModule[
-        TransactionRepository & TransactionProcessingStateRepository & 
-        TransactionManagerService & TransactionProcessor
+        TransactionRepository & TransactionProcessingStateRepository &
+            TransactionManagerService & TransactionProcessor
     ]
     with ScalatagsSupport:
